ingestion/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
qdrant-client==1.7.0
sentence-transformers==2.3.1
torch==2.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
boto3==1.29.0
